Change directory into redux
